pyproject.toml
[tool.poetry]
name = "citydashboard"
version = "0.2.0"
description = "CityDashboard — Panel de Control Municipal modular para ciudades españolas"
authors = ["PepeluiMoreno"]
readme = "README.md"
packages = [{ include = "app" }, { include = "models" }, { include = "graphql" }]

[tool.poetry.dependencies]
python = "^3.12"

# API
fastapi = "^0.115"
uvicorn = { version = "^0.32", extras = ["standard"] }
pydantic = "^2.10"
pydantic-settings = "^2.7"
jinja2 = "^3.1"
python-multipart = "^0.0.20"   # necesario para Form() en FastAPI

# GraphQL
strawberry-graphql = { version = "^0.249", extras = ["fastapi"] }

# Base de datos
sqlalchemy = { version = "^2.0", extras = ["asyncio"] }
asyncpg = "^0.30"
alembic = "^1.14"
greenlet = "^3.1"

# ETL (Sprint 02 — se declaran ya para no tocar pyproject después)
httpx = "^0.28"
openpyxl = "^3.1"
xlrd = "^2.0"          # lectura de .xls legacy (2023 y anteriores)
pandas = "^2.2"
pdfplumber = "^0.11"
beautifulsoup4 = "^4.12"
lxml = "^5.3"

# Tareas asíncronas (Sprint 02)
celery = { version = "^5.4", extras = ["redis"] }
redis = "^5.2"

# Dashboard (Sprint 07)
dash = "^2.18"
plotly = "^5.24"

# Storage
boto3 = "^1.35"          # MinIO compatible con S3 API

# Utilidades
python-dateutil = "^2.9"
structlog = "^24.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.24"
pytest-cov = "^6.0"
httpx = "^0.28"          # TestClient async
factory-boy = "^3.3"
faker = "^33.0"
ruff = "^0.8"
mypy = "^1.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
